feat: #747 MyProfileViewModel과 MyBooksViewModel 통합
