+++ android/TodokTodok/app/src/androidTest/java/com/team/todoktodok/compose/my/component/ProfileTabTest.kt
@@ -4,6 +4,7 @@ import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import com.team.todoktodok.presentation.compose.my.component.ProfileTab
+import com.team.todoktodok.presentation.compose.my.model.MyProfileUiState
 import org.junit.Rule
 import org.junit.Test
 
@@ -17,6 +18,7 @@ class ProfileTabTest {
         composeTestRule.setContent {
             ProfileTab(
                 onChangeBottomNavigationTab = {},
+                uiState = MyProfileUiState(),
             )
         }
