+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/presentation/compose/my/books/ActivatedBooksScreen.kt
@@ -9,7 +9,6 @@ import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
@@ -17,50 +16,18 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavHostController
 import coil3.compose.SubcomposeAsyncImage
 import coil3.request.ImageRequest
 import coil3.request.crossfade
 import com.team.domain.model.Book
-import com.team.todoktodok.App
 import com.team.todoktodok.R
 import com.team.todoktodok.presentation.compose.core.component.ResourceNotFoundContent
 import com.team.todoktodok.presentation.compose.main.MainDestination
-import com.team.todoktodok.presentation.compose.my.books.vm.MyBooksViewModel
-import com.team.todoktodok.presentation.compose.my.books.vm.MyBooksViewModelFactory
 import com.team.todoktodok.presentation.compose.my.component.BookErrorIllustration
 import com.team.todoktodok.presentation.compose.my.component.ShimmerPlaceholder
 import com.team.todoktodok.presentation.compose.preview.MyBooksUiStatePreviewParameterProvider
 import com.team.todoktodok.presentation.compose.theme.GrayE0
-
-@Composable
-fun ActivatedBooksScreen(
-    navController: NavHostController,
-    onChangeBottomNavigationTab: (MainDestination) -> Unit,
-    modifier: Modifier = Modifier,
-    viewModel: MyBooksViewModel =
-        viewModel(
-            factory =
-                MyBooksViewModelFactory(
-                    (LocalContext.current.applicationContext as App).container,
-                ),
-        ),
-) {
-    val uiState = viewModel.uiState.collectAsStateWithLifecycle()
-
-    LaunchedEffect(Unit) {
-        viewModel.loadMyBooks()
-    }
-
-    ActivatedBooksScreen(
-        uiState = uiState.value,
-        navController = navController,
-        onChangeBottomNavigationTab = onChangeBottomNavigationTab,
-        modifier = modifier,
-    )
-}
 
 @Composable
 fun ActivatedBooksScreen(
@@ -131,7 +98,6 @@ private fun ActivatedBooksScreenPreview(
     uiState: MyBooksUiState,
 ) {
     ActivatedBooksScreen(
-        navController = NavHostController(LocalContext.current),
         onChangeBottomNavigationTab = {},
         uiState = uiState,
     )
@@ -141,7 +107,6 @@ private fun ActivatedBooksScreenPreview(
 @Composable
 private fun EmptyActivatedBooksScreenPreview() {
     ActivatedBooksScreen(
-        navController = NavHostController(LocalContext.current),
         onChangeBottomNavigationTab = {},
         uiState = MyBooksUiState(),
     )
