--- android/TodokTodok/app/src/main/java/com/team/todoktodok/presentation/compose/my/books/MyBooksUiEvent.kt
@@ -1,9 +1,0 @@
-package com.team.todoktodok.presentation.compose.my.books
-
-import com.team.domain.model.exception.TodokTodokExceptions
-
-sealed interface MyBooksUiEvent {
-    data class ShowErrorMessage(
-        val exceptions: TodokTodokExceptions,
-    ) : MyBooksUiEvent
-}
