+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/presentation/compose/my/component/ProfileTab.kt
@@ -24,6 +24,7 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import com.team.todoktodok.R
@@ -31,7 +32,9 @@ import com.team.todoktodok.presentation.compose.discussion.model.DiscussionTabSt
 import com.team.todoktodok.presentation.compose.main.MainDestination
 import com.team.todoktodok.presentation.compose.my.books.ActivatedBooksScreen
 import com.team.todoktodok.presentation.compose.my.liked.LikedDiscussionsScreen
+import com.team.todoktodok.presentation.compose.my.model.MyProfileUiState
 import com.team.todoktodok.presentation.compose.my.participated.ParticipatedDiscussionsScreen
+import com.team.todoktodok.presentation.compose.preview.MyProfileUiStatePreviewParameterProvider
 import com.team.todoktodok.presentation.compose.theme.Green1A
 import com.team.todoktodok.presentation.compose.theme.Pretendard
 import com.team.todoktodok.presentation.compose.theme.White
@@ -56,6 +59,7 @@ enum class ProfileTabDestination(
 
 @Composable
 fun ProfileTab(
+    uiState: MyProfileUiState,
     onChangeBottomNavigationTab: (MainDestination) -> Unit,
     modifier: Modifier = Modifier,
     navController: NavHostController = NavHostController(LocalContext.current),
@@ -68,6 +72,7 @@ fun ProfileTab(
     Column(modifier = modifier) {
         ProfileTabRow(pagerState)
         ProfileTabPager(
+            uiState = uiState,
             navController = navController,
             onChangeBottomNavigationTab = onChangeBottomNavigationTab,
             pagerState = pagerState,
@@ -115,6 +120,7 @@ private fun ProfileTabRow(pagerState: PagerState) {
 
 @Composable
 private fun ProfileTabPager(
+    uiState: MyProfileUiState,
     navController: NavHostController,
     onChangeBottomNavigationTab: (MainDestination) -> Unit,
     pagerState: PagerState,
@@ -133,6 +139,7 @@ private fun ProfileTabPager(
             when (ProfileTabDestination.entries[page]) {
                 ProfileTabDestination.ACTIVATED_BOOKS ->
                     ActivatedBooksScreen(
+                        uiState = uiState.activatedBooks,
                         navController = navController,
                         onChangeBottomNavigationTab = onChangeBottomNavigationTab,
                     )
@@ -146,8 +153,12 @@ private fun ProfileTabPager(
 
 @Preview
 @Composable
-private fun ProfileTabPreview() {
+private fun ProfileTabPreview(
+    @PreviewParameter(MyProfileUiStatePreviewParameterProvider::class)
+    uiState: MyProfileUiState,
+) {
     ProfileTab(
+        uiState = uiState,
         navController = NavHostController(LocalContext.current),
         onChangeBottomNavigationTab = {},
     )
