+++ android/TodokTodok/app/src/test/java/com/team/todoktodok/presentation/vm/MyProfileViewModelTest.kt
@@ -2,7 +2,9 @@ package com.team.todoktodok.presentation.vm
 
 import app.cash.turbine.test
 import com.team.domain.ConnectivityObserver
+import com.team.domain.model.Book
 import com.team.domain.model.exception.NetworkResult
+import com.team.domain.model.member.MemberId
 import com.team.domain.model.member.Profile
 import com.team.domain.repository.MemberRepository
 import com.team.todoktodok.CoroutinesTestExtension
@@ -56,6 +58,27 @@ class MyProfileViewModelTest {
             viewModel.uiState.test {
                 val item = awaitItem()
                 assertEquals(profile, item.profile)
+            }
+        }
+
+    @Test
+    fun `내가 활동한 책 목록을 불러온다`() =
+        runTest {
+            // Given
+            val books =
+                listOf(
+                    Book(1, "Book 1", "페토", ""),
+                    Book(2, "Book 2", "정페토", ""),
+                )
+            coEvery { memberRepository.getMemberBooks(MemberId.Mine) } returns NetworkResult.Success(books)
+
+            // When
+            viewModel.loadMyBooks()
+
+            // Then
+            viewModel.uiState.test {
+                val item = awaitItem()
+                assertEquals(books, item.activatedBooks.books)
             }
         }
 }
